Fit a Bradley--Terry model

Fits Bradley--Terry model, also known as a quasi-symmetry model, to paired comparison data via maximum (quasi) likelihood estimation.

BradleyTerry(C, family = stats::quasibinomial)

Arguments

C
a square matrix of paired comparisons

family
a description of the error distribution and link function to be used. See family

Value

An object of class "glm"

Details

This implementation avoids calling BradleyTerry2::BTm() and instead fits the glm directly. By default, the model is fitted by quasi-likelihood (allowing non-integer counts) with a logit link. To fit by maximum likelihood instead, set the family argument to binomial.

References

Bradley, R. A., & Terry, M. E. (1952). Rank analysis of incomplete block designs: I. The method of paired comparisons. Biometrika, 39(3/4), 324--345.

See also

Other network centrality estimators: BTscores, ILSR, PageRank, Scroogefactor

Other Bradley-Terry model utility functions: hessianBT, loglikelihood, pseudocount, vcovBT

Examples

BradleyTerry(citations)

#> 
#> Call:  glm(formula = Y ~ -1 + X, family = family)
#> 
#> Coefficients:
#>    XAISM      XAoS     XANZS     XBern     XBioJ      XBcs      XBka    XBiost  
#> -0.08609   0.97383  -0.34590   0.29011  -0.80365   0.44562   0.88551   0.25394  
#>     XCJS     XCSSC     XCSTM    XCmpSt     XCSDA      XEES     XEnvr      XISR  
#> -0.09811  -1.66296  -1.14724  -1.04148  -0.92679  -0.88309  -0.51053  -0.07490  
#>   XJABES     XJASA      XJAS      XJBS     XJCGS      XJMA      XJNS   XJRSS-A  
#> -0.56758   0.85904  -1.81551  -1.23721   0.24181  -0.85459  -0.92963   0.29315  
#>  XJRSS-B   XJRSS-C     XJSCS     XJSPI      XJSS     XJTSA      XLDA     XMtka  
#>  1.68822  -0.31563  -1.31890  -0.73115  -1.20490  -0.03200  -0.30345  -0.58642  
#>     XSJS   XStataJ    XStCmp    XStats    XStMed     XSMMR    XStMod    XStNee  
#>  0.25756  -0.37883  -0.36682  -1.05620  -0.33798  -0.75304  -0.62272  -0.50782  
#>   XStPap      XSPL    XStSci    XStSin     XTech     XTest  
#> -1.74933  -0.49789  -0.28948  -0.11411   0.12942  -1.10290  
#> 
#> Degrees of Freedom: 871 Total (i.e. Null);  825 Residual
#> Null Deviance:	    5092 
#> Residual Deviance: 1688 	AIC: NA