Estimate the parameters of a Bradley-Terry model

Estimate Bradley-Terry ability scores from paired comparison data using the I-LSR algorithm. Avoids using stats::glm() and seems to be faster, but only approximates maximum likelihood.

ILSR(C, sort = FALSE, maxits = 100, tolerance = 1e-06, verbose = FALSE)

Arguments

C
a square matrix of paired comparisons

sort
logical. If TRUE, sort the weights in descending order

maxits
The maximum number of iterations in the I-LSR algorithm

tolerance
The criterion for convergence of the algorithm

verbose
logical. If TRUE, show the number of iterations as a message

Value

A vector of estimated ability scores

Details

The function uses iterative Luce Spectral Ranking (I-LSR) to approximate the maximum likelihood ability weights of the Bradley-Terry model. To do: Add unit tests.

References

Maystre, Lucas and Grossglauser, Matthias (2015). Fast and accurate inference of Plackett--Luce models. In Advances in Neural Information Processing Systems, 172--180.

See also

Other network centrality estimators: BTscores, BradleyTerry, PageRank, Scroogefactor

Examples

ILSR(citations)

#>         AmS        AISM         AoS        ANZS        Bern        BioJ 
#> 0.023845803 0.021878755 0.063145181 0.016872929 0.031871582 0.010675591 
#>         Bcs         Bka       Biost         CJS        CSSC        CSTM 
#> 0.037234178 0.057807370 0.030739517 0.021617390 0.004520609 0.007571317 
#>       CmpSt        CSDA         EES        Envr         ISR       JABES 
#> 0.008415962 0.009438678 0.009860333 0.014311652 0.022124895 0.013518074 
#>        JASA         JAS         JBS        JCGS         JMA         JNS 
#> 0.056297501 0.003881006 0.006919858 0.030368921 0.010145372 0.009411896 
#>      JRSS-A      JRSS-B      JRSS-C        JSCS        JSPI         JSS 
#> 0.031968723 0.129001550 0.017391450 0.006377069 0.011478352 0.007147140 
#>        JTSA         LDA        Mtka         SJS      StataJ       StCmp 
#> 0.023094738 0.017604588 0.013265793 0.030851089 0.016326278 0.016523634 
#>       Stats       StMed        SMMR       StMod       StNee       StPap 
#> 0.008292948 0.017007064 0.011229792 0.012792923 0.014350610 0.004146546 
#>         SPL       StSci       StSin        Tech        Test 
#> 0.014493825 0.017852214 0.021274276 0.027140416 0.007914610