Compute a PageRank vector

Given a (weighted) adjacency matrix, compute the PageRank: the the stationary distribution of a random walk around the graph.

PageRank(C, alpha = 0.85, sort = FALSE)

Arguments

C
a square matrix

alpha
a damping factor

sort
logical. Reorder the indices in descending order of PageRank

Value

A PageRank vector, scaled to sum to one

Details

PageRank is an eigenvector centrality metric, equivalent to Pinski & Narin's influence weight (1976) with the addition of a damping factor alpha, which simulates a random surfer traversing the graph and teleporting at any time with probability 1 - alpha. The effect of the damping factor is to smooth out any disconnected components or transient portions of the network.

In bibliometrics, PageRank has also been implemented as the Eigenfactor Metric and as the SCImago Journal Rank.

By default, C[i,j] refers to the directed edge that points from column j to row i. Use t(C) if you want edges directed from rows to columns instead.

References

Pinski, G., & Narin, F. (1976). Citation influence for journal aggregates of scientific publications: Theory, with application to the literature of physics. Information Processing & Management, 12(5), 297--312.

Page, L., Brin, S., Motwani, R., & Winograd, T. (1999). The PageRank citation ranking: bringing order to the web. Technical Report, Stanford InfoLab.

See also

Other network centrality estimators: BTscores, BradleyTerry, ILSR, Scroogefactor

Examples

PageRank(citations)

#>         AmS        AISM         AoS        ANZS        Bern        BioJ 
#> 0.012202138 0.010862596 0.098915149 0.006697535 0.015483828 0.012464668 
#>         Bcs         Bka       Biost         CJS        CSSC        CSTM 
#> 0.059136945 0.058378620 0.023982870 0.012621280 0.007618752 0.014087489 
#>       CmpSt        CSDA         EES        Envr         ISR       JABES 
#> 0.006069933 0.039376887 0.007432336 0.009385664 0.008349565 0.006919707 
#>        JASA         JAS         JBS        JCGS         JMA         JNS 
#> 0.110303393 0.006908512 0.008526762 0.022145121 0.027459149 0.009329186 
#>      JRSS-A      JRSS-B      JRSS-C        JSCS        JSPI         JSS 
#> 0.012862422 0.067875638 0.011081492 0.008798648 0.037841892 0.009230503 
#>        JTSA         LDA        Mtka         SJS      StataJ       StCmp 
#> 0.008850493 0.009731145 0.007994149 0.019257984 0.010430958 0.012695927 
#>       Stats       StMed        SMMR       StMod       StNee       StPap 
#> 0.006542398 0.061516023 0.008532377 0.006185163 0.005717473 0.005254096 
#>         SPL       StSci       StSin        Tech        Test 
#> 0.024401487 0.020397719 0.027139599 0.013314648 0.009689682 

# Scroogefactor, an estimator for the Bradley-Terry model
PageRank(citations)/colSums(citations)

#>          AmS         AISM          AoS         ANZS         Bern         BioJ 
#> 7.395235e-05 6.664169e-05 1.231820e-04 6.765187e-05 7.741914e-05 2.960729e-05 
#>          Bcs          Bka        Biost          CJS         CSSC         CSTM 
#> 7.457370e-05 1.172261e-04 6.680465e-05 5.897795e-05 1.822668e-05 2.194313e-05 
#>        CmpSt         CSDA          EES         Envr          ISR        JABES 
#> 4.895107e-05 2.297368e-05 5.546519e-05 5.272845e-05 1.043696e-04 5.766422e-05 
#>         JASA          JAS          JBS         JCGS          JMA          JNS 
#> 1.107464e-04 1.775967e-05 2.292140e-05 7.260695e-05 2.597838e-05 3.205906e-05 
#>       JRSS-A       JRSS-B       JRSS-C         JSCS         JSPI          JSS 
#> 1.004877e-04 2.620681e-04 5.083253e-05 2.802117e-05 2.832477e-05 3.205036e-05 
#>         JTSA          LDA         Mtka          SJS       StataJ        StCmp 
#> 8.349522e-05 5.094840e-05 4.844939e-05 7.860402e-05 9.070398e-05 4.755029e-05 
#>        Stats        StMed         SMMR        StMod        StNee        StPap 
#> 4.450611e-05 3.676989e-05 3.440475e-05 5.425582e-05 7.425289e-05 2.871090e-05 
#>          SPL        StSci        StSin         Tech         Test 
#> 3.714077e-05 4.879837e-05 4.441833e-05 7.197107e-05 3.166563e-05