Compute the Scroogefactor

The Scroogefactor is PageRank divided by out-degree. It can be used as an approximate estimator for the Bradley--Terry model.

Scroogefactor(C, alpha = 1, sort = FALSE)

Arguments

C
a square matrix

alpha
a damping factor

sort
logical. Reorder the indices in descending order of Scroogefactor score

Value

A vector equivalent to PageRank per reference, scaled to sum to one

Details

Pinksi and Narin (1976) first proposed this metric as a citation metric called the influence per reference. When applied to citation data, it can be interpreted as the influence per outgoing reference, effectively penalising larger publications and review journals, which have larger combined bibliographies. When applied to sports or games results, it is influence per game lost.

References

Pinski, G., & Narin, F. (1976). Citation influence for journal aggregates of scientific publications: Theory, with application to the literature of physics. Information Processing & Management, 12(5), 297--312.

See also

Other network centrality estimators: BTscores, BradleyTerry, ILSR, PageRank

Examples

Scroogefactor(citations, alpha = 1, sort = TRUE)

#>      JRSS-B         AoS         Bka        JASA      JRSS-A         Bcs 
#> 0.129637467 0.064100143 0.056911430 0.055443856 0.035544777 0.034351333 
#>         SJS        JCGS        Bern       Biost         ISR        Tech 
#> 0.032038128 0.031520003 0.030353184 0.028950225 0.025075321 0.024375400 
#>         AmS         CJS        AISM       StSci        JTSA       StSin 
#> 0.023797146 0.021343516 0.020799389 0.020455001 0.019693464 0.019512690 
#>       StCmp      JRSS-C        ANZS      StataJ         LDA       StMed 
#> 0.018277029 0.017204101 0.016219058 0.016153895 0.015975592 0.015942379 
#>        Envr       StNee         SPL       JABES        JSPI         EES 
#> 0.015481649 0.014689126 0.014201014 0.013103574 0.011927699 0.011867634 
#>       StMod        Mtka       CmpSt        Test         JMA       Stats 
#> 0.011629628 0.011399224 0.011348236 0.011155904 0.011075746 0.010253172 
#>        BioJ         JNS        CSDA        SMMR         JSS        JSCS 
#> 0.010023670 0.009670011 0.009646062 0.009208731 0.008576811 0.007460998 
#>        CSTM         JBS        CSSC       StPap         JAS 
#> 0.006206337 0.005827482 0.004046142 0.003833824 0.003692799