Calculate citation profiles

For a directed network, calculate the outgoing transition probabilities of a random walker at each node. This is equivalent to the proportion of citations each journal allocates to each other journal.

cprofile(x, self)

# S3 method for default
cprofile(x, self = TRUE)

# S3 method for igraph
cprofile(x, self = TRUE)

Arguments

x
a matrix of citations (from columns to rows) or an igraph object

self
logical. Include self-citations? If FALSE, they will not be counted.

Value

A matrix of journal citation profiles.

See also

community_profile()

Examples

cprofile(citations)

#> 47 x 47 sparse Matrix of class "dgCMatrix"

#>    [[ suppressing 33 column names ‘AmS’, ‘AISM’, ‘AoS’ ... ]]

#>    [[ suppressing 33 column names ‘AmS’, ‘AISM’, ‘AoS’ ... ]]

#>                                                                                  
#> AmS   0.260606061 0.006134969 0.00249066 .          .     .           0.008827238
#> AISM  .           0.110429448 0.00373599 0.03030303 0.025 .           .          
#> AoS   0.054545455 0.147239264 0.36239103 0.04040404 0.265 0.042755344 0.036569987
#> ANZS  .           0.030674847 0.00249066 0.05050505 .     .           0.002522068
#> Bern  0.006060606 0.042944785 0.03362391 .          0.110 .           0.002522068
#> BioJ  .           .           0.00373599 0.02020202 .     0.130641330 0.018915511
#> Bcs   0.054545455 0.006134969 0.01867995 0.10101010 0.020 0.106888361 0.240857503
#> Bka   0.042424242 0.061349693 0.04732254 0.05050505 0.070 0.033254157 0.075662043
#> Biost 0.024242424 .           0.00747198 .          .     0.040380048 0.060529634
#> CJS   .           0.012269939 0.00996264 0.01010101 0.015 0.002375297 0.007566204
#> CSSC  0.012121212 0.030674847 .          0.02020202 .     0.011876485 0.002522068
#> CSTM  0.006060606 0.030674847 .          0.03030303 0.005 0.007125891 .          
#> CmpSt .           .           0.00124533 .          0.010 .           .          
#> CSDA  0.036363636 0.036809816 0.02366127 .          0.015 0.035629454 0.018915511
#> EES   .           .           .          .          .     .           0.005044136
#>                                                                              
#> AmS   0.002008032 0.008356546 0.009345794 0.011961722 0.004672897 0.016129032
#> AISM  0.006024096 .           .           0.014354067 0.028037383 .          
#> AoS   0.148594378 0.033426184 0.070093458 0.045454545 0.060747664 0.064516129
#> ANZS  0.004016064 .           .           0.002392344 0.004672897 0.008064516
#> Bern  0.008032129 .           0.009345794 0.004784689 0.014018692 0.024193548
#> BioJ  0.014056225 0.002785515 .           0.009569378 0.009345794 0.016129032
#> Bcs   0.062248996 0.161559889 0.079439252 0.026315789 0.024922118 0.016129032
#> Bka   0.150602410 0.072423398 0.074766355 0.026315789 0.037383178 0.032258065
#> Biost 0.030120482 0.150417827 0.014018692 0.007177033 0.006230530 .          
#> CJS   0.008032129 0.005571031 0.074766355 0.009569378 0.012461059 0.016129032
#> CSSC  .           .           .           0.069377990 0.029595016 0.040322581
#> CSTM  .           .           .           0.071770335 0.147975078 0.024193548
#> CmpSt 0.002008032 0.002785515 .           .           0.003115265 0.024193548
#> CSDA  0.020080321 0.016713092 0.051401869 0.098086124 0.038940810 0.177419355
#> EES   .           .           .           .           0.001557632 .          
#>                                                                         
#> AmS   0.008751459 0.014925373 .           0.0625 0.008333333 0.012048193
#> AISM  0.007001167 .           0.005617978 .      .           0.005020080
#> AoS   0.042590432 0.014925373 0.033707865 .      0.016666667 0.126506024
#> ANZS  0.004667445 .           .           .      .           0.005020080
#> Bern  0.005834306 .           .           .      .           0.011044177
#> BioJ  0.015169195 .           0.005617978 .      0.008333333 0.005020080
#> Bcs   0.036172695 0.029850746 0.044943820 0.0250 0.250000000 0.069277108
#> Bka   0.037339557 0.007462687 0.044943820 0.0625 0.041666667 0.093373494
#> Biost 0.007584597 0.014925373 .           .      .           0.033132530
#> CJS   0.011668611 0.014925373 0.011235955 .      .           0.014056225
#> CSSC  0.008751459 .           .           .      .           0.003012048
#> CSTM  0.015169195 0.007462687 0.016853933 0.0250 .           .          
#> CmpSt 0.007001167 .           .           .      0.016666667 0.004016064
#> CSDA  0.283547258 0.097014925 0.067415730 0.0500 0.075000000 0.022088353
#> EES   0.001750292 0.291044776 0.039325843 .      0.050000000 0.008032129
#>                                                                            
#> AmS   0.012853470 0.010752688 0.009836066 0.002838221 .           0.0078125
#> AISM  .           .           0.003278689 0.019867550 0.041237113 .        
#> AoS   0.023136247 0.021505376 0.101639344 0.128666036 0.154639175 .        
#> ANZS  0.012853470 .           0.003278689 0.011352886 0.003436426 .        
#> Bern  .           .           0.013114754 0.022705771 0.024054983 .        
#> BioJ  0.020565553 0.083333333 0.006557377 .           0.006872852 .        
#> Bcs   0.051413882 0.120967742 0.022950820 0.016083254 0.051546392 0.0312500
#> Bka   0.025706941 0.034946237 0.075409836 0.059602649 0.054982818 0.0078125
#> Biost 0.010282776 0.032258065 0.009836066 0.001892148 0.010309278 0.0703125
#> CJS   0.015424165 .           0.003278689 0.024597919 0.030927835 .        
#> CSSC  0.046272494 .           .           0.003784295 .           .        
#> CSTM  0.074550129 0.005376344 .           0.026490066 0.017182131 0.0078125
#> CmpSt 0.005141388 .           0.016393443 0.004730369 0.013745704 .        
#> CSDA  0.097686375 0.021505376 0.081967213 0.053926206 0.048109966 .        
#> EES   .           .           .           .           .           .        
#>                                                                              
#> AmS   0.003861004 0.018348624 0.003184713 0.008233533 0.027777778 0.009433962
#> AISM  0.011583012 .           0.012738854 0.020209581 0.003472222 0.028301887
#> AoS   0.146718147 0.032110092 0.050955414 0.080089820 0.024305556 0.084905660
#> ANZS  .           .           0.012738854 0.012724551 0.010416667 .          
#> Bern  0.011583012 .           0.012738854 0.014970060 0.006944444 0.037735849
#> BioJ  .           0.004587156 0.015923567 0.007485030 0.006944444 .          
#> Bcs   0.019305019 0.137614679 0.019108280 0.026946108 0.045138889 0.028301887
#> Bka   0.081081081 0.036697248 0.022292994 0.059880240 0.024305556 0.018867925
#> Biost .           0.059633028 0.015923567 0.004491018 0.020833333 .          
#> CJS   0.011583012 0.004587156 0.006369427 0.011976048 0.013888889 .          
#> CSSC  .           .           0.031847134 0.005239521 .           0.018867925
#> CSTM  0.003861004 0.004587156 0.073248408 0.028443114 .           .          
#> CmpSt 0.007722008 .           0.009554140 0.002245509 0.006944444 .          
#> CSDA  0.019305019 0.009174312 0.089171975 0.037425150 0.100694444 0.037735849
#> EES   .           0.013761468 0.003184713 0.002245509 .           .          
#>                                    
#> AmS   .          0.012121212 ......
#> AISM  .          0.054545455 ......
#> AoS   0.03141361 0.060606061 ......
#> ANZS  .          .           ......
#> Bern  .          0.006060606 ......
#> BioJ  0.03141361 0.012121212 ......
#> Bcs   0.14136126 0.018181818 ......
#> Bka   0.05235602 0.054545455 ......
#> Biost 0.05759162 .           ......
#> CJS   0.01570681 0.018181818 ......
#> CSSC  .          0.006060606 ......
#> CSTM  0.01047120 0.054545455 ......
#> CmpSt .          .           ......
#> CSDA  0.02094241 0.018181818 ......
#> EES   .          .           ......
#> 
#>  ..............................
#>  ........suppressing columns and rows in show(); maybe adjust 'options(max.print= *, width = *)'
#>  ..............................
#> 

#>    [[ suppressing 33 column names ‘AmS’, ‘AISM’, ‘AoS’ ... ]]

#>                                                                       
#> Mtka   .           0.012269939 0.00124533 .          .     .          
#> SJS    0.006060606 0.030674847 0.01867995 0.02020202 0.040 0.021377672
#> StataJ .           .           .          .          .     .          
#> StCmp  0.012121212 .           0.00622665 .          .     0.004750594
#> Stats  .           .           0.00373599 .          .     .          
#> StMed  0.115151515 0.006134969 0.00622665 0.01010101 .     0.244655582
#> SMMR   0.012121212 .           .          0.01010101 .     0.016627078
#> StMod  .           .           .          .          .     0.007125891
#> StNee  0.012121212 .           0.00622665 .          .     0.002375297
#> StPap  .           .           .          0.02020202 0.005 0.002375297
#> SPL    .           0.049079755 0.01494396 0.05050505 0.040 0.014251781
#> StSci  0.078787879 .           0.01618929 0.01010101 0.035 0.007125891
#> StSin  0.018181818 0.036809816 0.03362391 0.04040404 0.040 0.011876485
#> Tech   0.012121212 0.006134969 0.00373599 0.03030303 0.005 0.002375297
#> Test   .           .           0.01494396 .          .     .          
#>                                                                               
#> Mtka   .           0.004016064 .           .           0.009569378 0.015576324
#> SJS    0.012610340 0.026104418 0.036211699 0.009345794 0.004784689 0.007788162
#> StataJ .           .           .           .           .           .          
#> StCmp  0.013871375 0.008032129 0.019498607 0.014018692 0.004784689 0.001557632
#> Stats  0.001261034 0.004016064 .           .           0.011961722 0.014018692
#> StMed  0.093316520 0.040160643 0.086350975 0.060747664 0.088516746 0.034267913
#> SMMR   0.006305170 0.004016064 0.013927577 0.014018692 .           .          
#> StMod  0.003783102 .           0.005571031 0.009345794 0.009569378 .          
#> StNee  .           .           0.002785515 .           0.007177033 0.003115265
#> StPap  .           .           .           .           0.016746411 0.015576324
#> SPL    0.008827238 0.012048193 0.011142061 0.032710280 0.050239234 0.074766355
#> StSci  0.016393443 0.026104418 0.016713092 0.028037383 0.009569378 0.012461059
#> StSin  0.025220681 0.048192771 0.022284123 0.032710280 0.023923445 0.026479751
#> Tech   0.005044136 0.004016064 0.002785515 0.004672897 0.031100478 0.029595016
#> Test   0.005044136 0.008032129 .           .           0.007177033 0.007788162
#>                                                                         
#> Mtka   .          0.005250875 .           .           .      .          
#> SJS    0.02419355 0.016336056 .           0.028089888 .      0.008333333
#> StataJ .          0.003500583 .           .           .      .          
#> StCmp  0.03225806 0.014002334 .           0.005617978 0.0125 .          
#> Stats  .          0.004084014 .           0.005617978 .      .          
#> StMed  0.05645161 0.047257876 0.164179104 0.033707865 0.1000 0.033333333
#> SMMR   .          0.004084014 0.022388060 .           .      .          
#> StMod  .          0.004667445 .           0.016853933 .      .          
#> StNee  .          0.004084014 .           .           .      .          
#> StPap  .          0.006417736 .           .           .      .          
#> SPL    0.02419355 0.019253209 .           0.033707865 0.0750 .          
#> StSci  0.01612903 0.009918320 0.022388060 0.005617978 0.0125 0.016666667
#> StSin  0.02419355 0.024504084 0.007462687 0.011235955 0.0375 0.008333333
#> Tech   0.02419355 0.019836639 .           .           .      0.008333333
#> Test   .          0.006417736 .           0.011235955 .      .          
#>                                                                                
#> Mtka   0.001004016 0.015424165 .           .           0.0141911069 0.003436426
#> SJS    0.016064257 0.010282776 .           0.013114754 0.0227057711 0.030927835
#> StataJ .           .           .           .           .            .          
#> StCmp  0.013052209 0.010282776 .           0.039344262 0.0037842952 0.003436426
#> Stats  0.002008032 0.007712082 .           0.003278689 0.0122989593 0.003436426
#> StMed  0.027108434 0.089974293 0.241935484 0.003278689 0.0047303690 0.003436426
#> SMMR   0.001004016 0.012853470 0.005376344 .           0.0028382214 .          
#> StMod  0.002008032 0.017994859 .           0.009836066 0.0028382214 .          
#> StNee  0.001004016 0.010282776 .           .           0.0009460738 0.010309278
#> StPap  .           0.007712082 .           .           0.0037842952 .          
#> SPL    0.006024096 0.020565553 .           0.003278689 0.0586565752 0.068728522
#> StSci  0.018072289 0.012853470 0.018817204 0.019672131 0.0104068117 0.003436426
#> StSin  0.021084337 0.020565553 0.013440860 0.026229508 0.0293282876 0.037800687
#> Tech   0.009036145 0.023136247 0.010752688 0.019672131 0.0085146641 0.017182131
#> Test   0.010040161 0.002570694 0.005376344 .           0.0132450331 0.013745704
#>                                                                             
#> Mtka   .         .           .           0.025477707 0.024700599 .          
#> SJS    .         0.023166023 0.022935780 0.006369427 0.018712575 0.017361111
#> StataJ 0.0468750 .           .           .           0.001497006 .          
#> StCmp  0.0078125 0.030888031 0.013761468 0.003184713 0.003742515 0.003472222
#> Stats  .         .           .           0.019108280 0.014221557 .          
#> StMed  0.2031250 0.011583012 0.128440367 0.047770701 0.024700599 0.079861111
#> SMMR   0.0312500 .           0.027522936 .           0.002245509 0.006944444
#> StMod  .         .           .           .           0.003742515 0.013888889
#> StNee  0.0156250 .           0.009174312 0.009554140 0.002245509 .          
#> StPap  .         .           .           0.012738854 0.005988024 .          
#> SPL    .         0.011583012 0.027522936 0.047770701 0.057634731 0.003472222
#> StSci  0.0078125 0.038610039 0.022935780 0.028662420 0.008982036 0.017361111
#> StSin  .         0.034749035 0.027522936 0.015923567 0.036676647 0.013888889
#> Tech   .         0.003861004 0.004587156 0.015923567 0.026197605 0.017361111
#> Test   .         0.019305019 .           0.019108280 0.016467066 0.006944444
#>                                                  
#> Mtka   0.018867925 .           0.121212121 ......
#> SJS    0.009433962 0.089005236 0.024242424 ......
#> StataJ .           .           .           ......
#> StCmp  0.018867925 .           .           ......
#> Stats  .           .           .           ......
#> StMed  .           0.078534031 0.012121212 ......
#> SMMR   .           .           .           ......
#> StMod  .           .           .           ......
#> StNee  .           0.005235602 0.012121212 ......
#> StPap  .           .           0.024242424 ......
#> SPL    0.075471698 0.005235602 0.072727273 ......
#> StSci  0.047169811 0.015706806 0.006060606 ......
#> StSin  0.047169811 0.047120419 0.030303030 ......
#> Tech   .           0.010471204 0.036363636 ......
#> Test   .           .           .           ......

cprofile(citations, self = FALSE)

#> 47 x 47 sparse Matrix of class "dgCMatrix"

#>    [[ suppressing 33 column names ‘AmS’, ‘AISM’, ‘AoS’ ... ]]

#>    [[ suppressing 33 column names ‘AmS’, ‘AISM’, ‘AoS’ ... ]]

#>                                                                                         
#> AmS   0.000000000 0.006896552 0.003906250 .          .           .           0.011627907
#> AISM  .           0.000000000 0.005859375 0.03191489 0.028089888 .           .          
#> AoS   0.073770492 0.165517241 0.000000000 0.04255319 0.297752809 0.049180328 0.048172757
#> ANZS  .           0.034482759 0.003906250 0.00000000 .           .           0.003322259
#> Bern  0.008196721 0.048275862 0.052734375 .          0.000000000 .           0.003322259
#> BioJ  .           .           0.005859375 0.02127660 .           0.000000000 0.024916944
#> Bcs   0.073770492 0.006896552 0.029296875 0.10638298 0.022471910 0.122950820 0.000000000
#> Bka   0.057377049 0.068965517 0.074218750 0.05319149 0.078651685 0.038251366 0.099667774
#> Biost 0.032786885 .           0.011718750 .          .           0.046448087 0.079734219
#> CJS   .           0.013793103 0.015625000 0.01063830 0.016853933 0.002732240 0.009966777
#> CSSC  0.016393443 0.034482759 .           0.02127660 .           0.013661202 0.003322259
#> CSTM  0.008196721 0.034482759 .           0.03191489 0.005617978 0.008196721 .          
#> CmpSt .           .           0.001953125 .          0.011235955 .           .          
#> CSDA  0.049180328 0.041379310 0.037109375 .          0.016853933 0.040983607 0.024916944
#> EES   .           .           .           .          .           .           0.006644518
#>                                                                                         
#> AmS   0.002364066 0.009836066 0.01010101 0.012853470 0.005484461 0.016528926 0.012214984
#> AISM  0.007092199 .           .          0.015424165 0.032906764 .           0.009771987
#> AoS   0.174940898 0.039344262 0.07575758 0.048843188 0.071297989 0.066115702 0.059446254
#> ANZS  0.004728132 .           .          0.002570694 0.005484461 0.008264463 0.006514658
#> Bern  0.009456265 .           0.01010101 0.005141388 0.016453382 0.024793388 0.008143322
#> BioJ  0.016548463 0.003278689 .          0.010282776 0.010968921 0.016528926 0.021172638
#> Bcs   0.073286052 0.190163934 0.08585859 0.028277635 0.029250457 0.016528926 0.050488599
#> Bka   0.000000000 0.085245902 0.08080808 0.028277635 0.043875686 0.033057851 0.052117264
#> Biost 0.035460993 0.000000000 0.01515152 0.007712082 0.007312614 .           0.010586319
#> CJS   0.009456265 0.006557377 0.00000000 0.010282776 0.014625229 0.016528926 0.016286645
#> CSSC  .           .           .          0.000000000 0.034734918 0.041322314 0.012214984
#> CSTM  .           .           .          0.077120823 0.000000000 0.024793388 0.021172638
#> CmpSt 0.002364066 0.003278689 .          .           0.003656307 0.000000000 0.009771987
#> CSDA  0.023640662 0.019672131 0.05555556 0.105398458 0.045703839 0.181818182 0.000000000
#> EES   .           .           .          .           0.001828154 .           0.002442997
#>                                                                                        
#> AmS   0.02105263 .           0.08333333 0.009615385 0.015706806 0.014124294 0.014545455
#> AISM  .          0.006802721 .          .           0.006544503 .           .          
#> AoS   0.02105263 0.040816327 .          0.019230769 0.164921466 0.025423729 0.029090909
#> ANZS  .          .           .          .           0.006544503 0.014124294 .          
#> Bern  .          .           .          .           0.014397906 .           .          
#> BioJ  .          0.006802721 .          0.009615385 0.006544503 0.022598870 0.112727273
#> Bcs   0.04210526 0.054421769 0.03333333 0.288461538 0.090314136 0.056497175 0.163636364
#> Bka   0.01052632 0.054421769 0.08333333 0.048076923 0.121727749 0.028248588 0.047272727
#> Biost 0.02105263 .           .          .           0.043193717 0.011299435 0.043636364
#> CJS   0.02105263 0.013605442 .          .           0.018324607 0.016949153 .          
#> CSSC  .          .           .          .           0.003926702 0.050847458 .          
#> CSTM  0.01052632 0.020408163 0.03333333 .           .           0.081920904 0.007272727
#> CmpSt .          .           .          0.019230769 0.005235602 0.005649718 .          
#> CSDA  0.13684211 0.081632653 0.06666667 0.086538462 0.028795812 0.107344633 0.029090909
#> EES   0.00000000 0.047619048 .          0.057692308 0.010471204 .           .          
#>                                                                                         
#> AmS   0.011494253 0.003444317 .           0.01123596 0.004901961 0.019801980 0.003558719
#> AISM  0.003831418 0.024110218 0.043636364 .          0.014705882 .           0.014234875
#> AoS   0.118773946 0.156142365 0.163636364 .          0.186274510 0.034653465 0.056939502
#> ANZS  0.003831418 0.013777268 0.003636364 .          .           .           0.014234875
#> Bern  0.015325670 0.027554535 0.025454545 .          0.014705882 .           0.014234875
#> BioJ  0.007662835 .           0.007272727 .          .           0.004950495 0.017793594
#> Bcs   0.026819923 0.019517796 0.054545455 0.04494382 0.024509804 0.148514851 0.021352313
#> Bka   0.088122605 0.072330654 0.058181818 0.01123596 0.102941176 0.039603960 0.024911032
#> Biost 0.011494253 0.002296211 0.010909091 0.10112360 .           0.064356436 0.017793594
#> CJS   0.003831418 0.029850746 0.032727273 .          0.014705882 0.004950495 0.007117438
#> CSSC  .           0.004592423 .           .          .           .           0.035587189
#> CSTM  .           0.032146958 0.018181818 0.01123596 0.004901961 0.004950495 0.081850534
#> CmpSt 0.019157088 0.005740528 0.014545455 .          0.009803922 .           0.010676157
#> CSDA  0.095785441 0.065442021 0.050909091 .          0.024509804 0.009900990 0.099644128
#> EES   .           .           .           .          .           0.014851485 0.003558719
#>                                                                       
#> AmS   0.009936766 0.040609137 0.01234568 .          0.013793103 ......
#> AISM  0.024390244 0.005076142 0.03703704 .          0.062068966 ......
#> AoS   0.096657633 0.035532995 0.11111111 0.03529412 0.068965517 ......
#> ANZS  0.015356820 0.015228426 .          .          .           ......
#> Bern  0.018066847 0.010152284 0.04938272 .          0.006896552 ......
#> BioJ  0.009033424 0.010152284 .          0.03529412 0.013793103 ......
#> Bcs   0.032520325 0.065989848 0.03703704 0.15882353 0.020689655 ......
#> Bka   0.072267389 0.035532995 0.02469136 0.05882353 0.062068966 ......
#> Biost 0.005420054 0.030456853 .          0.06470588 .           ......
#> CJS   0.014453478 0.020304569 .          0.01764706 0.020689655 ......
#> CSSC  0.006323397 .           0.02469136 .          0.006896552 ......
#> CSTM  0.034327010 .           .          0.01176471 0.062068966 ......
#> CmpSt 0.002710027 0.010152284 .          .          .           ......
#> CSDA  0.045167118 0.147208122 0.04938272 0.02352941 0.020689655 ......
#> EES   0.002710027 .           .          .          .           ......
#> 
#>  ..............................
#>  ........suppressing columns and rows in show(); maybe adjust 'options(max.print= *, width = *)'
#>  ..............................
#> 

#>    [[ suppressing 33 column names ‘AmS’, ‘AISM’, ‘AoS’ ... ]]

#>                                                                              
#> Mtka   .           0.013793103 0.001953125 .          .           .          
#> SJS    0.008196721 0.034482759 0.029296875 0.02127660 0.044943820 0.024590164
#> StataJ .           .           .           .          .           .          
#> StCmp  0.016393443 .           0.009765625 .          .           0.005464481
#> Stats  .           .           0.005859375 .          .           .          
#> StMed  0.155737705 0.006896552 0.009765625 0.01063830 .           0.281420765
#> SMMR   0.016393443 .           .           0.01063830 .           0.019125683
#> StMod  .           .           .           .          .           0.008196721
#> StNee  0.016393443 .           0.009765625 .          .           0.002732240
#> StPap  .           .           .           0.02127660 0.005617978 0.002732240
#> SPL    .           0.055172414 0.023437500 0.05319149 0.044943820 0.016393443
#> StSci  0.106557377 .           0.025390625 0.01063830 0.039325843 0.008196721
#> StSin  0.024590164 0.041379310 0.052734375 0.04255319 0.044943820 0.013661202
#> Tech   0.016393443 0.006896552 0.005859375 0.03191489 0.005617978 0.002732240
#> Test   .           .           0.023437500 .          .           .          
#>                                                                               
#> Mtka   .           0.004728132 .           .           0.010282776 0.018281536
#> SJS    0.016611296 0.030732861 0.042622951 0.010101010 0.005141388 0.009140768
#> StataJ .           .           .           .           .           .          
#> StCmp  0.018272425 0.009456265 0.022950820 0.015151515 0.005141388 0.001828154
#> Stats  0.001661130 0.004728132 .           .           0.012853470 0.016453382
#> StMed  0.122923588 0.047281324 0.101639344 0.065656566 0.095115681 0.040219378
#> SMMR   0.008305648 0.004728132 0.016393443 0.015151515 .           .          
#> StMod  0.004983389 .           0.006557377 0.010101010 0.010282776 .          
#> StNee  .           .           0.003278689 .           0.007712082 0.003656307
#> StPap  .           .           .           .           0.017994859 0.018281536
#> SPL    0.011627907 0.014184397 0.013114754 0.035353535 0.053984576 0.087751371
#> StSci  0.021594684 0.030732861 0.019672131 0.030303030 0.010282776 0.014625229
#> StSin  0.033222591 0.056737589 0.026229508 0.035353535 0.025706941 0.031078611
#> Tech   0.006644518 0.004728132 0.003278689 0.005050505 0.033419023 0.034734918
#> Test   0.006644518 0.009456265 .           .           0.007712082 0.009140768
#>                                                                                        
#> Mtka   .          0.007328990 .          .           .          .           0.001308901
#> SJS    0.02479339 0.022801303 .          0.034013605 .          0.009615385 0.020942408
#> StataJ .          0.004885993 .          .           .          .           .          
#> StCmp  0.03305785 0.019543974 .          0.006802721 0.01666667 .           0.017015707
#> Stats  .          0.005700326 .          0.006802721 .          .           0.002617801
#> StMed  0.05785124 0.065960912 0.23157895 0.040816327 0.13333333 0.038461538 0.035340314
#> SMMR   .          0.005700326 0.03157895 .           .          .           0.001308901
#> StMod  .          0.006514658 .          0.020408163 .          .           0.002617801
#> StNee  .          0.005700326 .          .           .          .           0.001308901
#> StPap  .          0.008957655 .          .           .          .           .          
#> SPL    0.02479339 0.026872964 .          0.040816327 0.10000000 .           0.007853403
#> StSci  0.01652893 0.013843648 0.03157895 0.006802721 0.01666667 0.019230769 0.023560209
#> StSin  0.02479339 0.034201954 0.01052632 0.013605442 0.05000000 0.009615385 0.027486911
#> Tech   0.02479339 0.027687296 .          .           .          0.009615385 0.011780105
#> Test   .          0.008957655 .          0.013605442 .          .           0.013089005
#>                                                                              
#> Mtka   0.016949153 .           .           0.017221584 0.003636364 .         
#> SJS    0.011299435 .           0.015325670 0.027554535 0.032727273 .         
#> StataJ .           .           .           .           .           0.06741573
#> StCmp  0.011299435 .           0.045977011 0.004592423 0.003636364 0.01123596
#> Stats  0.008474576 .           0.003831418 0.014925373 0.003636364 .         
#> StMed  0.098870056 0.327272727 0.003831418 0.005740528 0.003636364 0.29213483
#> SMMR   0.014124294 0.007272727 .           0.003444317 .           0.04494382
#> StMod  0.019774011 .           0.011494253 0.003444317 .           .         
#> StNee  0.011299435 .           .           0.001148106 0.010909091 0.02247191
#> StPap  0.008474576 .           .           0.004592423 .           .         
#> SPL    0.022598870 .           0.003831418 0.071182549 0.072727273 .         
#> StSci  0.014124294 0.025454545 0.022988506 0.012629162 0.003636364 0.01123596
#> StSin  0.022598870 0.018181818 0.030651341 0.035591274 0.040000000 .         
#> Tech   0.025423729 0.014545455 0.022988506 0.010332951 0.018181818 .         
#> Test   0.002824859 0.007272727 .           0.016073479 0.014545455 .         
#>                                                                              
#> Mtka   .           .           0.028469751 0.029810298 .           0.02469136
#> SJS    0.029411765 0.024752475 0.007117438 0.022583559 0.025380711 0.01234568
#> StataJ .           .           .           0.001806685 .           .         
#> StCmp  0.039215686 0.014851485 0.003558719 0.004516712 0.005076142 0.02469136
#> Stats  .           .           0.021352313 0.017163505 .           .         
#> StMed  0.014705882 0.138613861 0.053380783 0.029810298 0.116751269 .         
#> SMMR   .           0.029702970 .           0.002710027 0.010152284 .         
#> StMod  .           .           .           0.004516712 0.020304569 .         
#> StNee  .           0.009900990 0.010676157 0.002710027 .           .         
#> StPap  .           .           0.014234875 0.007226739 .           .         
#> SPL    0.014705882 0.029702970 0.053380783 0.069557362 0.005076142 0.09876543
#> StSci  0.049019608 0.024752475 0.032028470 0.010840108 0.025380711 0.06172840
#> StSin  0.044117647 0.029702970 0.017793594 0.044263776 0.020304569 0.06172840
#> Tech   0.004901961 0.004950495 0.017793594 0.031616983 0.025380711 .         
#> Test   0.024509804 .           0.021352313 0.019873532 0.010152284 .         
#>                                      
#> Mtka   .           0.000000000 ......
#> SJS    0.100000000 0.027586207 ......
#> StataJ .           .           ......
#> StCmp  .           .           ......
#> Stats  .           .           ......
#> StMed  0.088235294 0.013793103 ......
#> SMMR   .           .           ......
#> StMod  .           .           ......
#> StNee  0.005882353 0.013793103 ......
#> StPap  .           0.027586207 ......
#> SPL    0.005882353 0.082758621 ......
#> StSci  0.017647059 0.006896552 ......
#> StSin  0.052941176 0.034482759 ......
#> Tech   0.011764706 0.041379310 ......
#> Test   .           .           ......