Predict citations based on community profile

Based on the nearest community profile, what is the expected distribution of citations from journal idx?

fitted_citations(idx = NULL, citations, communities, self = TRUE)

Arguments

idx
A journal name or index. Vectorised for nearest_profile()

citations
a matrix of citations (from columns to rows) or an igraph object

communities
A membership vector or igraph::communities object

self
logical. Include self-citations? If FALSE, they will not be counted.

Value

A vector of predicted outgoing citations from journal idx.

Details

To vectorise this function, it might be easier first to create a vectorised version of nearest_point().

See also

Other functions for residual analysis of communities: community_residuals, profile_residuals

Examples

distances <- as.dist(1 - cor(citations + t(citations) - diag(diag(citations))))
clusters <- cutree(hclust(distances), h = 0.6)
fitted_citations(NULL, citations, clusters)

#>                 AmS        AISM          AoS       ANZS        Bern       BioJ
#> AmS    3.232653e+01  1.03542057 3.799478e+00  0.6683883  0.94632077  3.4291315
#> AISM   8.592441e-18  3.40603950 8.909121e+00  1.4739298  2.21895905  0.4339986
#> AoS    6.061224e+00 18.85853746 1.363882e+02  8.5420320 33.96965247 10.7964948
#> ANZS   0.000000e+00  1.31257309 4.061511e+00  0.5558448  1.01158427  0.2599191
#> Bern   6.734694e-01  3.53661724 1.821129e+01  1.4871840  4.53581335  0.5019628
#> BioJ   0.000000e+00  0.99797282 3.668461e+00  1.0797329  0.91368902 14.6942697
#> Bcs    7.408163e+00  5.31774873 3.249209e+01  5.9586897  8.09267417 57.5239721
#> Bka    8.081633e+00 10.15989142 6.367401e+01  5.5014719 15.85903084 20.4925493
#> Biost  2.693878e+00  1.31883771 9.302170e+00  1.6415190  2.31685430 25.1295619
#> CJS    8.352776e-18  2.36547303 1.349470e+01  1.2004569  3.36107032  2.2332196
#> CSSC   1.346939e+00  1.47187286 1.965247e+00  0.6867017  0.48947626  2.0126201
#> CSTM   2.020408e+00  6.53741549 6.419807e+00  2.9437970  1.59895578  1.9374869
#> CmpSt  4.282109e-18  0.54194366 3.930494e+00  0.2976723  0.97895252  0.2599191
#> CSDA   6.734694e+00  5.63349563 2.908566e+01  3.8912889  7.24424865 12.9114729
#> EES    3.637356e-17  0.28361289 1.965247e+00  1.9009594  0.48947626  1.0063257
#> Envr   2.580313e-17  0.92668506 5.895742e+00  1.7622279  1.46842878  1.7992210
#> ISR    1.548980e+01  0.68717605 2.227280e+00  0.3950045  0.55473976  1.0730899
#> JABES  2.158049e-17  0.34729365 1.834231e+00  1.1404541  0.45684451  1.4653687
#> JASA   1.279592e+01 17.43630588 1.172597e+02 10.3707572 29.20541687 39.5745561
#> JAS    9.710159e-18  1.18350559 1.310165e+00  0.5386842  0.32631751  1.3592221
#> JBS    0.000000e+00  0.16385839 1.310165e-01  0.6785194  0.03263175 14.8469050
#>                Bcs        Bka      Biost        CJS       CSSC       CSTM
#> AmS      5.7013563  2.5671240  2.5776905  2.2402905  5.6834919  5.0091171
#> AISM     2.6441716  4.8826707  1.4162382  1.5570666  5.9322785 17.9271661
#> AoS     45.5715266 75.5298936 24.0784698 23.3616168 21.4378944 44.0735425
#> ANZS     1.1135462  2.2281576  0.6487089  0.7653450  2.9541483  6.6390153
#> Bern     4.5835035  9.9405388  2.6216056  2.9518926  4.5401604 12.8670619
#> BioJ    19.4665782  4.2178427  8.6262390  2.4941223  5.7972825  5.0154635
#> Bcs     88.9218998 26.3961696 38.5129433 14.5375722 18.6239920 16.9840302
#> Bka     42.6816067 37.6112160 20.3014148 13.4938276 17.1179380 30.3067015
#> Biost   33.9352044  8.8938880 15.2190195  4.5836477  5.8180038  3.4026021
#> CJS      6.5601921  7.6356078  3.1858907  2.7283208  4.5822836  8.0460176
#> CSSC     2.5030521  1.3190916  1.1658310  0.8234509  7.2803285  9.7104928
#> CSTM     4.1543758  3.6971263  1.8619380  1.8656490 17.6521067 42.0079776
#> CmpSt    1.5672934  2.1570983  0.7480236  0.8438914  1.6029078  1.4395458
#> CSDA    27.1953198 17.3070333 11.7039001 10.5880274 45.5443821 27.1224473
#> EES     14.2356558  1.2178106  4.0708235  3.4036465  0.5480552  0.7385299
#> Envr    12.9028177  3.4635298  4.1942443  3.4170279  1.8490730  2.7694870
#> ISR      2.2189402  1.3599291  0.9712633  1.0731311  3.1239467  3.4995138
#> JABES    8.9263070  1.2100520  2.7724325  2.0943597  1.3468075  1.3716450
#> JASA    85.7376893 69.5103259 39.6078954 26.3800370 32.3246869 47.6274223
#> JAS      1.5310937  0.8625143  0.7445401  0.6654700  6.8390438  8.1619688
#> JBS     18.6063068  2.3498830  8.2209626  1.7300467  2.6828485  1.1304284
#>             CmpSt       CSDA          EES        Envr         ISR        JABES
#> AmS     1.0374619  15.809796 9.309663e-01  1.23451995 11.84255250  0.856865441
#> AISM    0.9757741  11.993639 3.096422e-01  1.44301296  0.54370714  0.250820343
#> AoS     8.9036395  77.413486 3.102541e+00  7.58996767  3.55451726  2.903492452
#> ANZS    0.7538343  10.903308 2.867143e-18  0.42653418  0.20309654  0.011700137
#> Bern    1.1900270  10.358142 0.000000e+00  1.00286725  0.63409777  0.035248611
#> BioJ    1.5156338  24.532443 6.192844e-01  0.91930742  0.15782808  1.219943101
#> Bcs     4.3912251  58.332697 1.300701e+01 14.14566305  3.20982037 12.557701126
#> Bka     5.3584424  57.242366 4.337030e+00  6.89912622  3.85456393  4.277298332
#> Biost   1.0828290  13.629135 6.192844e-01  0.93009773  1.07979892  1.779440603
#> CJS     1.5919474  20.171120 1.240609e+00  1.85769555  0.25282578  1.025843104
#> CSSC    2.4516932  41.977735 4.024729e-18  0.53045280  0.79085202  0.088229179
#> CSTM    3.7146275  61.603690 1.238569e+00  3.61594136  2.01324156  0.985869856
#> CmpSt   0.7495111  10.903308 6.192844e-01  0.73741514  0.04634241  0.465272348
#> CSDA   20.1768047 338.002545 1.052987e+01 11.90977373  3.34757349  8.258973374
#> EES     0.1888110   2.180662 1.610139e+01 15.84437579  0.02285384 11.848936064
#> Envr    0.6594052   8.177481 1.207605e+01 12.07072951  0.08570190  8.942053493
#> ISR     0.6377283   9.812977 6.233640e-01  0.82796868  5.71561674  0.507387514
#> JABES   0.4943954   7.632316 8.669981e+00  8.59131400  0.04341247  6.426493142
#> JASA   11.3743118 131.930025 1.455318e+01 18.64834183  6.09579197 12.774521332
#> JAS     2.4605114  42.522901 0.000000e+00  0.43196678  0.25754491  0.053035040
#> JBS     0.2522493   4.361323 0.000000e+00  0.06082128  0.03525740  0.780993207
#>               JASA        JAS         JBS       JCGS         JMA         JNS
#> AmS      5.1155269  5.0158891  3.01519757  2.0561187   6.4578574  1.89399152
#> AISM     9.6980689  4.0441524  0.30151976  2.7595843  20.5358353  5.83768432
#> AoS    148.7077729 17.8705733  9.27173252 37.9854035 130.8615852 34.34909205
#> ANZS     4.3595228  2.2671648  0.15075988  1.5095483   8.0376408  2.29614233
#> Bern    19.4837625  3.1829288  0.37689970  5.0129317  23.0891904  6.28004167
#> BioJ     7.0204510  6.5890465 13.26686930  1.9728776   6.2255590  1.75575460
#> Bcs     51.6994821 21.5422693 52.23829787 13.9748235  35.7244703  9.68302592
#> Bka     73.7284600 15.6492428 18.31732523 19.3236483  68.5683631 18.31856778
#> Biost   15.0803365  7.4373548 22.91550152  3.2058061   9.1050764  2.46413307
#> CJS     15.3457577  3.9719534  1.88449848  4.5487331  15.7022357  4.26332265
#> CSSC     2.4153598  6.6872732  1.50759878  1.8724537   8.5028159  2.47389187
#> CSTM     7.6791607 13.5973843  1.28145897  3.9545526  37.3107105 10.93466061
#> CmpSt    4.5028634  1.5155419  0.15075988  1.5916638   3.7627910  1.00936947
#> CSDA    37.7443466 45.7351781  9.12097264 22.3480989  36.7940260 10.32118778
#> EES      9.6430421  0.5648001  0.90455927  4.6752258   1.9510947  0.51330440
#> Envr    12.1364186  1.7150765  1.58297872  4.9207050   6.2714790  1.68953238
#> ISR      2.8488915  2.7317430  0.90455927  1.0839571   4.2278697  1.19745424
#> JABES    6.1886150  1.3747544  1.28145897  2.9248128   2.2763896  0.61535189
#> JASA   139.0907494 30.8929135 35.20243161 38.1369347 119.2497482 31.63444939
#> JAS      1.5888086  6.3308487  0.90455927  1.8540571   6.7837305  2.01969122
#> JBS      3.0328063  3.7551397 13.56838906  0.2619265   0.9292900  0.29785946
#>             JRSS-A      JRSS-B      JRSS-C       JSCS        JSPI          JSS
#> AmS     2.75372899  1.22548540  2.53127382  2.9368398   9.9372897 8.000000e+00
#> AISM    0.26917814  2.87355197  0.68044117  5.4495768  36.2437757 1.000000e+00
#> AoS     3.57666869 43.99069995 11.77456935 17.6442100 104.6348597 7.000000e+00
#> ANZS    0.13984263  1.31000163  0.27465324  2.5393815  13.2546424 3.000000e+00
#> Bern    0.28672822  5.87387828  1.14698803  4.0637653  27.6786582 2.000000e+00
#> BioJ    3.62305507  1.18322728  5.04774605  3.6717866   9.5292722 2.000000e+00
#> Bcs    14.78795630 10.48001305 24.10021277 10.7730024  37.9273271 1.300000e+01
#> Bka     5.79818361 20.53744493 11.40791235 12.8760988  67.6112341 7.000000e+00
#> Biost   6.38451039  3.00032632  8.83539999  2.7411446   7.6388443 6.000000e+00
#> CJS     0.71623879  4.35258607  1.75270183  3.7104466  17.2466185 4.000000e+00
#> CSSC    0.56374674  0.63387176  0.67958906  5.8642664  17.7257025 0.000000e+00
#> CSTM    0.85749568  2.07064774  1.21676544 15.5030680  81.4688433 1.960602e-17
#> CmpSt   0.07610495  1.26774351  0.44400518  1.2416515   3.0745021 2.000000e+00
#> CSDA    3.43227693  9.38130201  8.00646679 34.3498957  45.5890176 2.900000e+01
#> EES     0.24927400  0.63387176  5.01468306  0.3902013   4.0848169 0.000000e+00
#> Envr    0.47367437  1.90161527  4.32072550  1.4024251   7.8274606 2.000000e+00
#> ISR     1.28563086  0.71838799  1.09452306  1.8676602   6.9208575 1.000000e+00
#> JABES   0.35544664  0.59161364  3.02224449  0.9995365   3.9721300 0.000000e+00
#> JASA   10.71274753 37.82101485 23.33402414 23.8037721 108.2788553 1.300000e+01
#> JAS     0.35045656  0.42258117  0.38996019  5.4889736  14.5514597 3.000000e+00
#> JBS     3.66921082  0.04225812  4.78442598  1.0892811   2.0703690 2.000000e+00
#>               JTSA        LDA        Mtka        SJS StataJ      StCmp
#> AmS     3.17189504  1.3086975  1.27358491  2.2245490      1  1.4959314
#> AISM    2.09063069  1.0552454  4.81132075  2.1608568      0  2.5537928
#> AoS     9.99474527 14.3377709 11.55660377 25.1192452      3 36.4306108
#> ANZS    0.80147715  0.4451083  1.74528302  0.9423269      0  1.2525617
#> Bern    2.00644328  1.7118094  3.44339623  3.3653998      0  4.7754506
#> BioJ    0.62346235  4.0529210  1.22641509  3.0244841      0  1.5965010
#> Bcs     3.47142425 18.7850703  4.29245283 16.7851192      1 12.2251169
#> Bka     6.02373931 11.0156741  7.92452830 15.0411846      1 18.1005933
#> Biost   0.89717793  7.1177140  0.84905660  5.2668619      0  2.7747520
#> CJS     1.29803318  1.8504471  2.07547170  3.0297477      1  4.0782691
#> CSSC    1.11753744  0.6790665  2.40566038  1.1700418      0  1.4080920
#> CSTM    4.49253370  1.5385948 11.08490566  3.3188790      0  3.2143340
#> CmpSt   0.28818563  0.4522379  0.33018868  0.8830913      0  1.3526077
#> CSDA    4.19214093  6.4674303  5.47169811 11.4951461      1 16.9049774
#> EES     0.14693998  2.5676459  0.18867925  3.9822409      0  3.9914268
#> Envr    0.49445822  2.5508301  0.70754717  3.9194635      0  4.2416750
#> ISR     1.64203305  0.5485280  0.89622642  1.1379863      2  0.9040529
#> JABES   0.20211132  1.6717541  0.33018868  2.4593805      0  2.4846971
#> JASA   10.12102706 21.4571041 12.21698113 29.2115399      0 35.3217936
#> JAS     0.83496084  0.4427023  1.98113208  0.9014051      0  1.2557955
#> JBS     0.11417935  3.7271139  0.28301887  2.1241475      0  0.1403010
#>             Stats       StMed        SMMR      StMod     StNee      StPap
#> AmS     1.3565621  13.5602837  2.01013171  1.4200066 0.5388469  1.4151047
#> AISM    3.9390064   1.3560284  0.20101317  0.5176853 1.1053149  5.2981575
#> AoS     9.8911768  41.6978723  6.18115502  5.3964127 6.4637296 12.7746343
#> ANZS    1.4770719   0.6780142  0.10050659  0.4624247 0.4275818  1.9284418
#> Bern    2.8365243   1.6950355  0.25126646  0.6455187 1.1330897  3.7935953
#> BioJ    1.1620821  59.6652482  8.84457953  2.3825850 1.2259041  1.3720136
#> Bcs     3.9237638 234.9319149 34.82553191  8.9073907 5.2421391  4.7744813
#> Bka     6.8270565  82.3787234 12.21155015  4.9690384 4.4160504  8.7639134
#> Biost   0.7973300 103.0581560 15.27700101  3.4780318 1.9805152  0.9464964
#> CJS     1.8204732   8.4751773  1.25633232  1.0934740 0.8705024  2.3004964
#> CSSC    2.2430110   6.7801418  1.00506586  1.3610872 0.5760541  2.6851059
#> CSTM    9.3382566   5.7631206  0.85430598  1.9441145 2.1921370 12.2405543
#> CmpSt   0.3444974   0.6780142  0.10050659  0.4346994 0.1779616  0.3736922
#> CSDA    6.9211004  41.0198582  6.08064843 11.7153109 2.3980174  6.3501352
#> EES     0.1670559   4.0680851  0.60303951  0.3232774 0.1497181  0.2094834
#> Envr    0.6264598   7.1191489  1.05531915  0.6515801 0.3966356  0.7855628
#> ISR     0.9119931   4.0680851  0.60303951  0.5955641 0.2858643  0.9945003
#> JABES   0.3204284   5.7631206  0.85430598  0.4557612 0.1964817  0.3704170
#> JASA   10.8317906 158.3163121 23.46828774 10.0822606 7.8153286 13.5547707
#> JAS     1.8985013   4.0680851  0.60303951  1.3732929 0.4433650  2.2192269
#> JBS     0.2586069  61.0212766  9.04559271  1.8180922 0.9839280  0.3153168
#>               SPL      StSci     StSin       Tech        Test
#> AmS     5.0711835 12.8460975   2.89101  3.2116104  1.79496822
#> AISM   19.1578045  1.5744700   6.77892  2.4299563  4.44499302
#> AoS    46.0162950 27.8331248 103.77729 16.0592363 39.03898595
#> ANZS    6.9493997  0.7943931   3.09039  1.1087668  1.82965710
#> Bern   13.7109777  3.1403046  13.85691  2.6758461  5.98560526
#> BioJ    4.8833619  8.1091981   2.79132  1.9800575  1.70362341
#> Bcs    17.0917667 39.6284648  24.72312  8.9193380 12.88943459
#> Bka    31.5540309 23.2293853  48.44934  9.9844625 20.11536813
#> Biost   3.3807890 15.2706075   7.07799  2.7988412  2.83875825
#> CJS     8.2641509  3.4632958  10.26807  2.2400614  4.56114042
#> CSSC    9.5789022  1.5394467   1.49535  1.9493837  1.78057643
#> CSTM   44.1380789  2.5705773   4.88481  5.4579571  7.14318481
#> CmpSt   1.3147513  0.8850132   2.99070  0.6641387  1.28937810
#> CSDA   21.7873070 15.1011296  22.13118 12.5332209 13.59584262
#> EES     0.7512864  4.6936588   1.49535  0.7745454  4.04499581
#> Envr    2.8173242  4.7435528   4.48605  1.2868182  4.39081326
#> ISR     3.5686106  5.8422640   1.69473  1.6628058  1.12791428
#> JABES   1.3147513  3.0797056   1.39566  0.7270736  2.50167562
#> JASA   48.6457976 44.5408679  89.22255 18.3451964 37.52867163
#> JAS     7.8885077  0.8128572   0.99690  1.7015735  1.44334927
#> JBS     1.1269297  7.6129668   0.09969  1.0048989  0.18833882
#>  [ reached getOption("max.print") -- omitted 26 rows ]

fitted_citations('Bka', citations, clusters)

#>               Bka
#> AmS     2.5671240
#> AISM    4.8826707
#> AoS    75.5298936
#> ANZS    2.2281576
#> Bern    9.9405388
#> BioJ    4.2178427
#> Bcs    26.3961696
#> Bka    37.6112160
#> Biost   8.8938880
#> CJS     7.6356078
#> CSSC    1.3190916
#> CSTM    3.6971263
#> CmpSt   2.1570983
#> CSDA   17.3070333
#> EES     1.2178106
#> Envr    3.4635298
#> ISR     1.3599291
#> JABES   1.2100520
#> JASA   69.5103259
#> JAS     0.8625143
#> JBS     2.3498830
#> JCGS   12.3471700
#> JMA    24.0396829
#> JNS     5.5348643
#> JRSS-A  3.0494408
#> JRSS-B 41.0375780
#> JRSS-C  3.8756000
#> JSCS    2.6516561
#> JSPI   18.0466160
#> JSS     2.0599061
#> JTSA    1.7891418
#> LDA     2.7562133
#> Mtka    2.2281576
#> SJS    10.8112308
#> StataJ  0.3291634
#> StCmp   6.2952848
#> Stats   2.2154974
#> StMed  22.2878625
#> SMMR    2.2200054
#> StMod   1.8263096
#> StNee   1.5604468
#> StPap   0.5100750
#> SPL    11.6831290
#> StSci   9.0796755
#> StSin  16.3256763
#> Tech    4.5624977
#> Test    4.5196158